overfix

+++ backend/mycloud/mycloud/views.py
@@ -1,19 +1,26 @@
-from django.core.exceptions import ObjectDoesNotExist
+from mimetypes import guess_type
+
+from django.contrib.auth.password_validation import validate_password
+from django.core.exceptions import ObjectDoesNotExist, ValidationError
 from rest_framework.authentication import TokenAuthentication
 from rest_framework.response import Response
 from rest_framework.decorators import action, api_view, permission_classes
 from rest_framework.permissions import IsAuthenticated, AllowAny
 from rest_framework.views import APIView
 from rest_framework import status, viewsets
+from rest_framework.authtoken.models import Token
 from django.utils.timezone import now
 from django.conf import settings
-from django.http import FileResponse, Http404
+from django.http import FileResponse, Http404, HttpResponse
 from django.views import View
 from datetime import timedelta
 from .models import File, TemporaryLink, CustomUser
 from .serializers import FileSerializer, UserSerializer
 import secrets
 import os
+from urllib.parse import quote
+import chardet
+
 
 
 class FileViewSet(viewsets.ModelViewSet):
@@ -23,7 +30,16 @@ class FileViewSet(viewsets.ModelViewSet):
 
     def get_queryset(self):
         """Фильтруем файлы по пользователю."""
-        return File.objects.filter(user=self.request.user)
+        user = self.request.user
+        user_id = self.request.query_params.get('user_id', None)
+
+        if (user.is_staff or user.is_superuser) and user_id:
+            return File.objects.filter(user_id=user_id)
+
+        if (user.is_staff or user.is_superuser):
+            return File.objects.all()
+
+        return File.objects.filter(user=user)
 
     def perform_create(self, serializer):
         """Передаем текущего пользователя в сериализатор."""
@@ -33,19 +49,35 @@ class FileViewSet(viewsets.ModelViewSet):
     def download(self, request, pk=None):
         """Функция для скачивания файла"""
         file = self.get_object()
-        if file.user != request.user:
+        if file.user != request.user and not request.user.is_staff:
             return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
 
         file_path = os.path.join(settings.MEDIA_ROOT, file.file.name)
-
-        # Проверяем, существует ли файл на сервере
         if not os.path.exists(file_path):
             return Response({"detail": "File not found"}, status=status.HTTP_404_NOT_FOUND)
 
-        # Отправляем файл
+        content_type, _ = guess_type(file_path)
+
+        if content_type and 'text' in content_type:
+            with open(file_path, 'rb') as f:
+                raw_data = f.read()
+                detected_encoding = chardet.detect(raw_data)['encoding']
+                if not detected_encoding:
+                    detected_encoding = 'utf-8'
+                decoded_data = raw_data.decode(detected_encoding).encode('utf-8')
+
+            response = HttpResponse(decoded_data, content_type='text/plain; charset=utf-8')
+            response['Content-Disposition'] = f'inline; filename="{file.name}"'
+            return response
+
         response = FileResponse(open(file_path, 'rb'))
-        response['Content-Type'] = 'application/octet-stream'
-        response['Content-Disposition'] = f'attachment; filename="{file.name}"'
+        if content_type:
+            response['Content-Type'] = content_type
+            response['Content-Disposition'] = f'inline; filename="{file.name}"'
+        else:
+            response['Content-Type'] = 'application/octet-stream'
+            response['Content-Disposition'] = f'attachment; filename="{file.name}"'
+
         return response
 
     @action(detail=True, methods=['delete'])
@@ -53,18 +85,14 @@ class FileViewSet(viewsets.ModelViewSet):
         """Удаление файла"""
         try:
             file = self.get_object()
-
-
-            if file.user != request.user:
+            if file.user != request.user and not request.user.is_staff:
                 return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
 
             file_path = file.file.path
-
             if os.path.exists(file_path):
                 os.remove(file_path)
 
             file.delete()
-
             return Response({"detail": "File deleted successfully"}, status=status.HTTP_204_NO_CONTENT)
 
         except ObjectDoesNotExist:
@@ -74,12 +102,13 @@ class FileViewSet(viewsets.ModelViewSet):
     def rename_file(self, request, pk=None):
         """Переименование файла"""
         file = self.get_object()
-        if file.user != request.user:
+        if file.user != request.user and not request.user.is_staff:
             return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
 
         new_name = request.data.get("name")
         if not new_name:
             return Response({"detail": "Name field is required"}, status=status.HTTP_400_BAD_REQUEST)
+
         file.name = new_name
         file.save()
         return Response({"detail": "File renamed successfully", "new_name": file.name}, status=status.HTTP_200_OK)
@@ -88,7 +117,7 @@ class FileViewSet(viewsets.ModelViewSet):
     def update_comment(self, request, pk=None):
         """Обновление комментария к файлу"""
         file = self.get_object()
-        if file.user != request.user:
+        if file.user != request.user and not request.user.is_staff:
             return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
 
         comment = request.data.get('comment')
@@ -105,14 +134,13 @@ class FileViewSet(viewsets.ModelViewSet):
     def generate_link(self, request, pk=None):
         """Генерация временной ссылки на файл"""
         file = self.get_object()
-        if file.user != request.user:
+        if file.user != request.user and not request.user.is_staff:
             return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
 
         token = secrets.token_urlsafe(32)
-        expires_at = now() + timedelta(hours=1)  # Ссылка будет действительна 1 час
+        expires_at = now() + timedelta(hours=1) # Ссылка будет действительна 1 час
 
         temporary_link = TemporaryLink.objects.create(file=file, token=token, expires_at=expires_at)
-
         link = f"{request.build_absolute_uri('/')[:-1]}/api/files/temp/{token}/"
         return Response({
             "link": link,
@@ -135,9 +163,13 @@ class TemporaryLinkDownloadView(View):
         if not os.path.exists(file_path):
             raise Http404("File not found")
 
+        # Для поддержки русских символов
+        filename = quote(temporary_link.file.name)
+
         response = FileResponse(open(file_path, 'rb'))
         response['Content-Type'] = 'application/octet-stream'
-        response['Content-Disposition'] = f'attachment; filename="{temporary_link.file.name}"'
+        response['Content-Disposition'] = f'attachment; filename="{filename}"; filename*=utf-8\'\'{filename}'
+
         return response
 
 
@@ -165,22 +197,26 @@ def list_users(request):
     return Response(serializer.data, status=status.HTTP_200_OK)
 
 
-@api_view(['DELETE'])
+@api_view(['GET', 'DELETE'])
 @permission_classes([IsAuthenticated])
-def delete_user(request, pk):
-    """Удаление пользователя для администратора"""
-    if not request.user.is_superuser:
-        return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
-
+def user_detail(request, user_id):
+    """Получение или удаление пользователя по ID"""
     try:
-        user = CustomUser.objects.get(pk=pk)
-        user.delete()
-        return Response({"detail": "User deleted successfully"}, status=status.HTTP_204_NO_CONTENT)
+        user = CustomUser.objects.get(id=user_id)
     except CustomUser.DoesNotExist:
         return Response({"detail": "User not found"}, status=status.HTTP_404_NOT_FOUND)
 
+    if request.method == 'GET':
+        if not (request.user.is_staff or request.user.is_superuser):
+            return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
+        serializer = UserSerializer(user)
+        return Response(serializer.data, status=status.HTTP_200_OK)
 
-from rest_framework.authtoken.models import Token
+    if request.method == 'DELETE':
+        if not (request.user.is_staff or request.user.is_superuser):
+            return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
+        user.delete()
+        return Response({"detail": "User deleted successfully"}, status=status.HTTP_204_NO_CONTENT)
 
 
 @api_view(['POST'])
@@ -194,19 +230,13 @@ def logout(request):
     return Response({"detail": "Logged out successfully"}, status=status.HTTP_200_OK)
 
 
-class UserProfileView(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def get(self, request):
-        """Получение атрибутов пользователя"""
-        user = request.user
-        data = {
-            "username": user.username,
-            "is_admin": user.is_staff,
-            "email": user.email
-        }
-        return Response(data)
-
+@api_view(['GET'])
+@permission_classes([IsAuthenticated])
+def current_user_profile(request):
+    """Получение атрибутов текущего пользователя"""
+    user = request.user
+    serializer = UserSerializer(user)
+    return Response(serializer.data, status=status.HTTP_200_OK)
 
 class TokenVerifyView(APIView):
     authentication_classes = [TokenAuthentication]
@@ -251,3 +281,92 @@ def summary(self, request):
     }
 
     return Response(summary_data, status=status.HTTP_200_OK)
+
+
+@api_view(['PATCH'])
+@permission_classes([IsAuthenticated])
+def update_user_profile(request):
+    """Редактирование данных профиля пользователя"""
+    user = request.user
+    serializer = UserSerializer(user, data=request.data, partial=True)
+    
+    if serializer.is_valid():
+        serializer.save()
+        return Response(serializer.data, status=status.HTTP_200_OK)
+    return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
+
+@api_view(['PATCH'])
+@permission_classes([IsAuthenticated])
+def update_user_by_id(request, user_id):
+    """Редактирование данных профиля пользователя (имя, фамилия, email)"""
+    try:
+        user = CustomUser.objects.get(id=user_id)
+    except CustomUser.DoesNotExist:
+        return Response({"detail": "User not found."}, status=status.HTTP_404_NOT_FOUND)
+
+    if 'first_name' in request.data:
+        user.first_name = request.data['first_name']
+    if 'last_name' in request.data:
+        user.last_name = request.data['last_name']
+    if 'email' in request.data:
+        user.email = request.data['email']
+
+    user.save()
+    serializer = UserSerializer(user)
+    return Response(serializer.data, status=status.HTTP_200_OK)
+
+
+@api_view(['POST'])
+@permission_classes([IsAuthenticated])
+def change_password(request, user_id):
+    """Изменение пароля пользователя по ID"""
+    if user_id:
+        if not request.user.is_superuser and request.user.id != user_id:
+            return Response({"detail": "Доступ запрещён"}, status=status.HTTP_403_FORBIDDEN)
+        try:
+            user = CustomUser.objects.get(id=user_id)
+        except CustomUser.DoesNotExist:
+            return Response({"detail": "User not found."}, status=status.HTTP_404_NOT_FOUND)
+    else:
+        user = request.user
+
+    new_password = request.data.get('new_password')
+    confirm_password = request.data.get('confirm_password')
+
+    if new_password != confirm_password:
+        return Response({"detail": "Пароли не совпадают"}, status=status.HTTP_400_BAD_REQUEST)
+
+    try:
+        validate_password(new_password)
+    except ValidationError as e:
+        return Response({"detail": str(e)}, status=status.HTTP_400_BAD_REQUEST)
+
+    user.set_password(new_password)
+    user.save()
+
+    return Response({"detail": "Пароль успешно изменён"}, status=status.HTTP_200_OK)
+
+
+
+@api_view(['PATCH'])
+@permission_classes([IsAuthenticated])
+def update_user_attributes(request, user_id):
+    """Редактирование атрибутов пользователя (is_active, is_staff, is_superuser)"""
+    if not (request.user.is_staff or request.user.is_superuser):
+        return Response({"detail": "Forbidden"}, status=status.HTTP_403_FORBIDDEN)
+
+    try:
+        user = CustomUser.objects.get(id=user_id)
+    except CustomUser.DoesNotExist:
+        return Response({"detail": "User not found"}, status=status.HTTP_404_NOT_FOUND)
+
+    allowed_fields = ['is_active', 'is_staff', 'is_superuser']
+    data = {key: value for key, value in request.data.items() if key in allowed_fields}
+
+    for field, value in data.items():
+        setattr(user, field, value)
+
+    user.save()
+
+    return Response(UserSerializer(user).data, status=status.HTTP_200_OK)
